Rename Logger to LoggerCollection for clarity
