+++ shoes-core/lib/shoes/logger_collection.rb
@@ -1,8 +1,7 @@
 # frozen_string_literal: true
-require 'logger'
 
 class Shoes
-  class Logger
+  class LoggerCollection
     def initialize
       @loggers = []
     end
@@ -28,19 +27,11 @@ class Shoes
       forward(:error, message)
     end
 
+    private
+
     def forward(meth, message)
       @loggers.each do |logger|
         logger.public_send(meth, message)
-      end
-    end
-
-    class StandardLogger < SimpleDelegator
-      def initialize(device = STDERR)
-        logger = ::Logger.new(device)
-        logger.formatter = proc do |severity, _datetime, _progname, message|
-          "#{severity}: #{message}\n"
-        end
-        super(logger)
       end
     end
   end
